Implement __repr__ for simplerandom.random RNGs.
The object could be represented either as a single (potentially
large) seed integer, or alternatively as a number of 32-bit seeds
(which is not the standard Python random API, but it's consistent
with the C init/seed API).
__repr__ shows the state as a single seed integer, for consistency
with the Python API.
Individual state integers are available via rng.getstate().